+++ client/src/components/admin/pumpsAdmin.js
@@ -1,0 +1,17 @@
+
+import React, {Component}  from 'react';
+import Client from '../../clients/Pump';
+
+class PumpsAdmin extends Component {
+
+  render = () => {
+    return (
+      <div className="admin-pumps">
+        Pump Admin
+      </div>    
+    )
+  }
+
+}
+
+export default PumpsAdmin;
