+++ client/src/index.js
@@ -7,6 +7,8 @@ import '../semantic/dist/semantic.min.css';
 import './components/drink/style.css';
 
 import Admin from './components/admin';
+import DrinksAdmin from './components/admin/drinksAdmin';
+import PumpsAdmin from './components/admin/pumpsAdmin';
 import Home from './components/home';
 
 
@@ -19,7 +21,9 @@ ReactDOM.render((
       <App>
         <Switch>
         <Route exact={true} path="/" component={Home}/>
-        <Route path="/admin" component={Admin}/>
+        <Route exact path="/admin" component={Admin}/>
+        <Route path="/admin/drinks" component={DrinksAdmin}/>
+        <Route path="/admin/pumps" component={PumpsAdmin}/>
         <Route path="*" component={NotFound}/>
         </Switch>
       </App>
